+++ Plugin.php
@@ -23,6 +23,8 @@ use Synder\Analytics\Widgets\SimpleSystems;
 
 class Plugin extends PluginBase
 {
+    const VERSION = '1.1.0';
+
     /**
      * Plugin dependencies
      *
@@ -57,13 +59,18 @@ class Plugin extends PluginBase
 
     /**
      * Boot Plugin
-     * @todo
      *
      * @return void
      */
     public function boot()
     {
-        if (Settings::get('dev_reevaluate') === 1) {
+        $version = Settings::get('version', '1.0.2');
+        if (version_compare($version, self::VERSION, "<")) {
+            $this->upgrade($version);
+        }
+
+        // Re-Evaluate Visitors
+        if (Settings::get('dev_reevaluate', 1) === 1) {
             $items = Visitor::all();
             foreach ($items AS $item) {
                 $item->evaluate(true);
@@ -71,6 +78,7 @@ class Plugin extends PluginBase
             Settings::set('dev_reevaluate', 0);
         }
 
+        // Extend Mohsin.Txt Plugin
         if (Settings::instance()->getRobotsTxtProvider() === 'mohsin.txt') {
             if (Settings::get('bot_robots') === '1') {
                 \Mohsin\Txt\Models\Robot::extend(function($model) {
@@ -83,12 +91,26 @@ class Plugin extends PluginBase
             }
         }
 
+        // Add Invisible Link
+        if (Settings::get('bot_inlink') === '1') {
+            Event::listen('cms.template.processTwigContent', function($object, &$dataHolder) {
+                $data = '<div style="top:0;color:transparent;position:absolute;text-indent:-9999999px;"><a href="{{ "'.Settings::get('bot_inlink_link').'"|app }}">Synder System</a></div>';
+
+                if (($offset = strpos($dataHolder->content, '</body>')) !== false) {
+                    $dataHolder->content = substr($dataHolder->content, 0, $offset) . $data . substr($dataHolder->content, $offset);
+                }
+            });
+        }
+
+        // Extend CMS Controller
         CmsController::extend(function($controller) {
             $controller->middleware(AnalyticsMiddleware::class);
         });
 
+        // Extend Page Model
         RequestPage::extend(fn($model) => $this->extendPageModel($model));
 
+        // Extend Backend
         //@todo
         //Event::listen('backend.menu.extendItems', function (NavigationManager $manager) {
         //    $manager->addSideMenuItem('October.Backend', 'dashboard', 'statistics', [
@@ -100,6 +122,41 @@ class Plugin extends PluginBase
         //        'order'       => 10
         //    ]);
         //});
+    }
+
+    /**
+     * Upgrade Plugin
+     *
+     * @return void
+     */
+    protected function upgrade($version)
+    {
+        if (\version_compare($version, '1.1.0', '<')) {
+            $instance = Settings::instance();
+            $instance->initSettingsData();
+            $instance->version = '1.1.0';
+            $instance->save();
+
+            $autoHide = [
+                '/favicon.ico',
+                '/robots.txt',
+                '/humans.txt',
+                '/sitemaps.xml',
+                '/sitemap.xml.gz',
+                '/sitemap.xml',
+                '/sitemap.txt',
+                '/sitemap_index.txt',
+                '/atom.xml',
+                '/rss.xml'
+            ];
+            foreach ($autoHide AS $hide) {
+                $page = Page::where('path', '=', $hide)->first();
+                if (!empty($page)) {
+                    $page->hide = 1;
+                    $page->save();
+                }
+            }
+        }
     }
 
     /**
@@ -203,5 +260,31 @@ class Plugin extends PluginBase
                 'syndervisits' => [Page::class, 'markupVisits'],
             ]
         ];
+    }
+
+    /**
+     * Register Scheduled Tasks
+     * 
+     * @return void
+     */
+    public function registerSchedule($schedule)
+    {
+        if (Settings::get('bot_robots') === '0' && Settings::get('bot_inlink') === '0') {
+            return;
+        }
+        if (Settings::get('bot_robots_relocate_cron') === '0' && Settings::get('bot_inlink_relocate_cron') === '0') {
+            return;
+        }
+
+        $schedule->call(function () {
+            $instance = \Synder\Analytics\Models\Settings::instance();
+
+            if (time() - $instance->get('bot_robots_time') > 90 * 24 * 60 * 60) {
+                $instance->set('bot_robots_link', '0');
+            }
+            if (time() - $instance->get('bot_inlink_time') > 90 * 24 * 60 * 60) {
+                $instance->set('bot_inlink_link', '0');
+            }
+        })->monthly();
     }
 }