+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 1.1.0
 -   Add: New settings page to configure the Synder Simple Analytics plugin.
 -   Add: New chart configuration properties (accessible on the widget itself).
 -   Add: New Bot probability feature to filter requests made by non-humans.
--   Add: Robots.txt and invisible link traps.
+-   Add: New Robots.txt and invisible link traps.
 -   Add: New 'synder-slider' formwidget, used to configure Bot probability filter.
 -   Add: New 'hide' column on `synder_analytics` to hide pages and those requests.
 -   Add: New 'browser' and 'os' columns on `synder_analytics_visitors` to skip evaluation on each (re-) load.
@@ -23,6 +23,8 @@ Version 1.1.0
 -   Add: robots.txt Support for [Arcane.Seo](https://octobercms.com/plugin/arcane-seo) plugin.
 -   Add: robots.txt Support for [Mohsin.Txt](https://octobercms.com/plugin/mohsin-txt) plugin.
 -   Add: robots.txt Support for [Zen.Robots](https://octobercms.com/plugin/zen-robots) plugin.
+-   Add: Scheduled Task for Robots and Invisible Link re-generation.
+-   Add: German Locale.
 -   Update: Don't skip favicon.ico, rather hide favicon.ico and robots.txt requests per default.
 -   Bugfix: Duplicated tick for the first counted date on basic statistics widget.
 -   Bugfix: Missing Date-Tick on basic statistics widget.